2007-05-31 17:30 UTC+0200 Przemyslaw Czerpak (druzus/at/priv.onet.pl)   * harbour/common.mak     ! added PCRE_LIB
  * harbour/bin/hb-func.sh
    ! fixed typo in last commit

  * harbour/source/hbpcre/pcredfa.c
  * harbour/source/hbpcre/pcreexec.c
  * harbour/source/hbpcre/pcrefind.c
  * harbour/source/hbpcre/pcreprni.c
  * harbour/source/hbpcre/pcrestud.c
  * harbour/source/hbpcre/pcretryf.c
    * cleaned some warnings

+++ harbour/source/hbpcre/pcretryf.c
@@ -98,7 +98,7 @@ EXPORT real_pcre *
 _pcre_try_flipped(const real_pcre *re, real_pcre *internal_re,
   const pcre_study_data *study, pcre_study_data *internal_study)
 {
-if (byteflip(re->magic_number, sizeof(re->magic_number)) != MAGIC_NUMBER)
+if ((unsigned long) byteflip(re->magic_number, sizeof(re->magic_number)) != MAGIC_NUMBER)
   return NULL;
 
 *internal_re = *re;           /* To copy other fields */
